+++ src/components/Header.jsx
@@ -5,7 +5,7 @@ import { styled } from "@mui/material/styles";
 import Box from '@mui/material/Box';
 import Paper from '@mui/material/Paper';
 import Grid from '@mui/material/Grid';
-import Logo from "./Logo.jsx";
+import Logo from "./logo/Logo.jsx";
 import MenuButtons from "./MenuButtons.jsx";
 import SearchBar from "./SearchBar.jsx";
 import AppBar from '@mui/material/AppBar';
